refactor(variant icons): convert to yaml
also:
- typescript
- rename "getVariantArray" to something more accurate
