+++ client/src/info_bubble/InfoBubbleControls/Variants.test.tsx
@@ -19,8 +19,8 @@ vi.mock('../../segments/info')
 
 // Provide mock variant icons so that we can test icons with `enabledWithFlag`
 vi.mock(
-  '../../segments/variant_icons.json',
-  async () => await import('../../segments/__mocks__/variant_icons.json')
+  '../../segments/variant_icons.yaml',
+  async () => await import('../../segments/__mocks__/variant_icons.yaml')
 )
 
 describe('Variants', () => {
