--- client/src/segments/__mocks__/variant_icons.json
@@ -1,30 +1,0 @@
-{
-  "building": {
-    "waterfront": {
-      "id": "waterfront",
-      "title": "Waterfront",
-      "color": "#ff0"
-    }
-  },
-  "direction": {
-    "inbound": {
-      "id": "direction-inbound",
-      "title": "Inbound"
-    },
-    "outbound": {
-      "id": "direction-outbound",
-      "title": "Outbound"
-    },
-    "empty": {
-      "id": "",
-      "title": "No one"
-    }
-  },
-  "flagged-variant": {
-    "foo": {
-      "id": "foo",
-      "title": "Flagged variant",
-      "enableWithFlag": "FLAGGED_VARIANT"
-    }
-  }
-}
